#10 qna update
