--- HappyhouseAPI/src/main/java/com/ssafy/happyhouse/qna/model/service/AnswerService.java
@@ -1,13 +1,0 @@
-package com.ssafy.happyhouse.qna.model.service;
-
-import com.ssafy.happyhouse.qna.model.Answer;
-
-public interface AnswerService {
-	public Answer detailAnswer(int parentNo);
-
-	public boolean writeAnswer(Answer answer);
-
-	public boolean updateAnswer(Answer answer);
-
-	public boolean deleteAnswer(int no);
-}
