+++ HappyhouseAPI/src/main/java/com/ssafy/happyhouse/qna/model/service/QnAService.java
@@ -2,9 +2,10 @@ package com.ssafy.happyhouse.qna.model.service;
 
 import java.util.List;
 
+import com.ssafy.happyhouse.qna.model.Answer;
 import com.ssafy.happyhouse.qna.model.Question;
 
-public interface QuestionService {
+public interface QnAService {
 	public List<Question> retrieveQuestion();
 
 	public Question detailQuestion(int no);
@@ -14,4 +15,12 @@ public interface QuestionService {
 	public boolean updateQuestion(Question question);
 
 	public boolean deleteQuestion(int no);
+	
+	public Answer detailAnswer(int parentNo);
+
+	public boolean writeAnswer(Answer answer);
+
+	public boolean updateAnswer(Answer answer);
+
+	public boolean deleteAnswer(int no);
 }
